split out context and scaled
